requirements.txt
pyTelegramBotAPI==4.14.0
supabase==2.3.0
razorpay==1.4.1
python-dotenv==1.0.0
flask==3.0.0
pytz==2024.1